backend/requirements.txt
fastapi==0.115.4
uvicorn==0.30.6
SQLAlchemy==2.0.36
pydantic==2.9.2
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose==3.3.0
email-validator==2.0.0.post2
python-multipart==0.0.9
psycopg2-binary==2.9.9
